% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactTable.R
\name{ComposeProb}
\alias{ComposeProb}
\alias{composeProb}
\title{Probability of a and b goals}
\usage{
composeProb(a, b)
}
\arguments{
\item{a}{Index of goals in home}

\item{b}{Index of gaols away}
}
\value{
The value of the index of a and b of the matrix
}
\description{
Calculate compose probability of the fact table
}
\examples{
composeProb(1, 1)
> 1
composeProb(2, 5)
> 3
}
\author{
Juan Ma
}
