% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactTable.R
\name{FTAGProb}
\alias{FTAGProb}
\title{Probability of a goals at home}
\usage{
FTAGProb(a)
}
\arguments{
\item{a}{Index of goals away}
}
\value{
The probability of get `a` goals away
}
\description{
Calculate the probability for getting `a` goals away
}
\examples{
FTAGProb(1)
> 1
FTAGProb(2)
> 3
}
\author{
Juan Ma
}
