% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactTable.R
\name{FTHGProb}
\alias{FTHGProb}
\title{Probability of a goals at home}
\usage{
FTHGProb(a)
}
\arguments{
\item{a}{Index of goals at home}
}
\value{
The probability of get `a` goals at home
}
\description{
Calculate the probability for getting `a` goals at home
}
\examples{
FTHGProb(1)
> 1
FTHGProb(2)
> 3
}
\author{
Juan Ma
}
