% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileMerger.R
\name{mergeFiles}
\alias{mergeFiles}
\title{Combine files}
\usage{
mergeFiles()
}
\value{
The result data frame with the merged files
}
\description{
Combine many files in a data frame with determined rows
}
\author{
Juan Ma
}
