% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ratio.R
\name{probRatio}
\alias{probRatio}
\title{Get the probability ratio}
\usage{
probRatio(md)
}
\arguments{
\item{md}{merged dataframe}
}
\value{
Returns a dataframe with home team goals, foreign team goals and its probability ratio.
}
\description{
Get the probability ratio between the joint probability and the product of 
each marginal probability.
}
\author{
Emiliano
}
