% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactTable.R
\name{setDataSet}
\alias{setDataSet}
\alias{setDataFrame}
\title{Set a data frame for analyzing}
\usage{
setDataFrame(a)
}
\arguments{
\item{a}{The new data frame}
}
\description{
Set a different data frame for calculate probability
}
\author{
Juan Ma
}
